requirements.txt
groq
chromadb
sentence-transformers
rank_bm25
flashrank
python-dotenv
streamlit
PyYAML
pypdf
upstash-redis
supabase
numpy
